Revert "refactor: cognito google signin 리팩토링"
This reverts commit 4685ee7b8f85aa051f09e246d9a7eb1412573fd2.
