+++ src/main/java/com/tradin/module/auth/controller/AuthController.java
@@ -1,14 +1,14 @@
 package com.tradin.module.auth.controller;
 
-import com.tradin.module.auth.controller.dto.request.TokenRequestDto;
 import com.tradin.module.auth.service.AuthService;
 import com.tradin.module.users.controller.dto.response.TokenResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -16,16 +16,9 @@ import javax.validation.Valid;
 public class AuthController {
     private final AuthService authService;
 
-    @Operation(summary = "인증 탭 - 구글 로그인&회원가입")
+    @Operation(summary = "인증 - 구글 로그인&회원가입")
     @PostMapping("/auth/cognito")
-    public ResponseEntity<String> googleAuth(@RequestParam String code) {
-        authService.requestTokenToCognito(code);
-        return ResponseEntity.ok("success");
-    }
-
-    @Operation(summary = "Cognito Redirect URL")
-    @PostMapping("/auth/token")
-    public ResponseEntity<TokenResponseDto> cognitoRedirect(@Valid @RequestBody TokenRequestDto request) {
-        return ResponseEntity.ok(authService.auth(request.toServiceDto()));
+    public ResponseEntity<TokenResponseDto> auth(@RequestParam String code) {
+        return ResponseEntity.ok(authService.auth(code));
     }
 }
