+++ src/main/java/com/tradin/module/feign/client/CognitoClient.java
@@ -1,6 +1,7 @@
 package com.tradin.module.feign.client;
 
 import com.tradin.module.feign.client.dto.cognito.AuthDto;
+import com.tradin.module.feign.client.dto.cognito.TokenDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -8,5 +9,5 @@ import org.springframework.web.bind.annotation.RequestBody;
 @FeignClient(name = "cognitoClient", url = "https://tradin.auth.ap-northeast-2.amazoncognito.com")
 public interface CognitoClient {
     @PostMapping("/oauth2/token")
-    void getAccessAndRefreshToken(@RequestBody AuthDto authDto);
+    TokenDto getAccessAndRefreshToken(@RequestBody AuthDto authDto);
 }
