+++ src/main/java/com/tradin/module/feign/service/CognitoFeignService.java
@@ -5,6 +5,7 @@ import com.tradin.module.feign.client.CognitoClient;
 import com.tradin.module.feign.client.CognitoJwkFeignClient;
 import com.tradin.module.feign.client.dto.cognito.AuthDto;
 import com.tradin.module.feign.client.dto.cognito.JwkDtos;
+import com.tradin.module.feign.client.dto.cognito.TokenDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,13 +18,13 @@ public class CognitoFeignService {
     private final CognitoClient cognitoClient;
     private final CognitoJwkFeignClient cognitoJwkFeignClient;
 
-    public void getTokenFromCognito(String code) {
+    public TokenDto getTokenFromCognito(String code) {
         final String COGNITO_CLIENT_ID = secretKeyManager.getCognitoClientId();
         final String COGNITO_AUTH_REDIRECT_URI = secretKeyManager.getCognitoAuthRedirectUri();
 
         AuthDto authDto = new AuthDto("authorization_code", COGNITO_CLIENT_ID, COGNITO_AUTH_REDIRECT_URI, code);
 
-        cognitoClient.getAccessAndRefreshToken(authDto);
+        return cognitoClient.getAccessAndRefreshToken(authDto);
     }
 
     public JwkDtos getJwkKey() {
